Fixing a few integer conversion related failures on non-windows platforms

+++ Gems/Atom/Feature/Common/Code/Source/SkinnedMesh/SkinnedMeshRenderProxy.h
@@ -45,8 +45,8 @@ namespace AZ
             void SetSkinningMatrices(const AZStd::vector<float>& data) override;
             void SetMorphTargetWeights(uint32_t lodIndex, const AZStd::vector<float>& weights) override;
 
-            size_t GetLodCount() const;
-            AZStd::span<const AZStd::unique_ptr<SkinnedMeshDispatchItem>> GetDispatchItems(size_t lodIndex) const;
+            uint32_t GetLodCount() const;
+            AZStd::span<const AZStd::unique_ptr<SkinnedMeshDispatchItem>> GetDispatchItems(uint32_t lodIndex) const;
         private:
 
             AZ_DISABLE_COPY_MOVE(SkinnedMeshRenderProxy);
